requirements.txt
pymupdf
deep-translator
python-pptx